Consolidate test models.
